feat(agents): Activity tab uses Live.tsx rich rendering via shared hook
Phase 3: useAgentActivity hook extracted from Live.tsx
- All event processing, state management, subscriptions in one hook
- Optional agentName param filters events to single agent
- Live.tsx now ~350 lines (was 765 after phase 1-2 split)

Phase 4: AgentToolStream component
- Calls useAgentActivity(agentName) for filtered events
- Renders AgentDrillDown from LiveRenderers.tsx
- Same rich tool trees, aggregation, drill-down as /live page

Phase 5: Wired into AgentDetail
- Activity tab now uses AgentToolStream instead of simple timeline
- Shares SSE connection via useWebSocket (no second EventSource)
